# docs/openapi-skeleton.yaml
openapi: 3.0.3
info:
  title: Shelfline
  version: 0.1.0
  description: >-
    Banking-as-a-platform product shelf for capital-aware deposit, lending, and
    payment SKUs distributed through partner channels with experience-ownership rules.
  x-product-slug: ai-fin-shelfline
  x-source: ai-in-financial/WEF_The_future__of_financial_services/
servers:
  - url: https://api.example.com
...
paths:
  /v1/skus:
    get:
      operationId: listProductSkus
      tags: [Skus]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, paused, retired]
      responses:
        "200":
          description: SKUs page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvelopeProductSku"
    post:
      operationId: createProductSku
      tags: [Skus]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductSkuCreate"
      responses:
        "201":
          description: SKU created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeProductSku"
  /v1/skus/{skuId}:
    parameters:
      - $ref: "#/components/parameters/SkuId"
    get:
      operationId: getProductSku
      tags: [Skus]
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      responses:
        "200":
          description: SKU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataEnvelopeProductSku"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    SkuId:
      name: skuId
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    ResponseMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: { type: string }
        timestamp: { type: string, format: date-time }
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
    ProductSku:
      type: object
      required: [id, name, productFamily, role, status]
      properties:
        id: { type: string }
        name: { type: string }
        productFamily:
          type: string
          enum: [deposit, lending, payment]
        role:
          type: string
          enum: [manufacturer, distributor, both]
        status:
          type: string
          enum: [draft, published, paused, retired]
        adjudicationMode:
          type: string
          enum: [bank, partner, dual]
        capitalConstraintId: { type: string }
        liquidityCapId: { type: string }
    ProductSkuCreate:
      type: object
      required: [name, productFamily, role]
      properties:
        name: { type: string }
        productFamily:
          type: string
          enum: [deposit, lending, payment]
        role:
          type: string
          enum: [manufacturer, distributor, both]
        adjudicationMode:
          type: string
          enum: [bank, partner, dual]
...
    DataEnvelopeProductSku:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/ProductSku"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ListEnvelopeProductSku:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/ProductSku"
            nextCursor: { type: string }
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...