# platform/api-server/src/domains/skus/handlers/sku.handlers.ts
/**
 * Handlers for Sku
 *
 * Generated by Handler Generator v2.0.0
 * Generator Version: 2.0.0
 * Domain: skus
 *
 * ⚠️  DO NOT EDIT THIS FILE MANUALLY
 * This file is auto-generated. Any manual changes will be overwritten.
 */

import type { FastifyRequest, FastifyReply } from "fastify";
import type { SkusDomainModule } from "../dependencies/skus-ddd.dependencies.js";
import type { CreateProductSkuInput, GetProductSkuInput, ListProductSkusInput, UpdateProductSkuInput } from "@shelfline/services/skus";
// Use cases are available via deps.useCases.<repoKey>.<verb>
// Execution context is automatically set up by route wrapper (wrapHandlerInExecutionContext)

/**
 * Handler for GET /v1/skus
 * Operation: listProductSkus
 */
export async function listProductSkus(
  request: FastifyRequest<{ Params: Record<string, string>; Querystring: { cursor?: string | string[], limit?: string | string[], status?: string | string[], productFamily?: string | string[] }; Body: unknown }>,
  reply: FastifyReply,
  deps: SkusDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: ListProductSkusInput = {
    ...(request.query ?? {}),
  };
  const result = await deps.useCases.skus.list.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for POST /v1/skus
 * Operation: createProductSku
 */
export async function createProductSku(
  request: FastifyRequest,
  reply: FastifyReply,
  deps: SkusDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: CreateProductSkuInput = {
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.skus.create.execute(input);
  return reply.code(201).send(result);
}


/**
 * Handler for GET /v1/skus/{skuId}
 * Operation: getProductSku
 */
export async function getProductSku(
  request: FastifyRequest<{ Params: { skuId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: SkusDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: GetProductSkuInput = {
    skuId: request.params.skuId,
  };
  const result = await deps.useCases.skus.get.execute(input);
  return reply.code(200).send(result);
}


/**
 * Handler for PATCH /v1/skus/{skuId}
 * Operation: updateProductSku
 */
export async function updateProductSku(
  request: FastifyRequest<{ Params: { skuId: string }; Querystring: Record<string, string | string[] | undefined>; Body: unknown }>,
  reply: FastifyReply,
  deps: SkusDomainModule
): Promise<void> {
  // Security: orgId validated by setOrgContext hook; use case gets it from execution context
  const input: UpdateProductSkuInput = {
    skuId: request.params.skuId,
    ...(request.body ?? {}),
  };
  const result = await deps.useCases.skus.update.execute(input);
  return reply.code(200).send(result);
}